Use 64-bit file manipulation API to support file systems with inode64 attributes, e.g., WSL.
- dreambbs.mk: Use `-D_FILE_OFFSET_BITS=64` option on 64-bit OSs
   for enabling 64-bit file manipulation API.
- maple/gem.c: `gem_state()`, maple/mail.c: `mbox_size()`,
   maple/post.c: `post_state()`, & so/mailgem.c: `mailgem_state()`:
      Adjust format string accordingly for relevant variables and values.

+++ maple/mail.c
@@ -2636,13 +2636,13 @@ mbox_size(
         outs(fpath);
 
         if (!stat(fpath, &st))
-            prints("\nTime: %s\nSize: %ld", Ctime(&st.st_mtime), st.st_size);
+            prints("\nTime: %s\nSize: %lld", Ctime(&st.st_mtime), (long long)st.st_size);
         vmsg(NULL);
     }
     else
     {
         stat(fpath, &st);
-        sprintf(buf, "容量大小: %d K", (int)st.st_size/1024);
+        sprintf(buf, "容量大小: %lld K", (long long)st.st_size/1024);
         vmsg(buf);
     }
