+++ so/mailgem.c
@@ -328,7 +328,7 @@ XO *xo)
     outs(fpath);
 
     if (!stat(fpath, &st))
-        prints("\nTime: %s\nSize: %ld", Ctime(&st.st_mtime), st.st_size);
+        prints("\nTime: %s\nSize: %lld", Ctime(&st.st_mtime), (long long)st.st_size);
 
     vmsg(NULL);
